splitting some things up to avoid Giant Dump of Code
