+++ dev/defaults.yml
@@ -29,14 +29,23 @@ projects:
       name: opscode-erchef
       cookie: erchef
       node: erchef@127.0.0.1
+    git:
+      uri: git://github.com/chef/oc_erchef.git
+      branch: master
   oc_bifrost: # TODO : unverified
     type: erlang
     database: opscode_bifrost
+    build:
+      # higher concurrency appears to cause eunit failures...
+      max_concurrency: 1
     service:
       rel-type: relx
       name:  oc_bifrost
       cookie: oc_bifrost
       node: oc_bifrost@127.0.0.1
+    git:
+      uri: https://github.com/chef/oc_bifrost.git
+      branch: master
   bookshelf: # TODO unverified
     type: erlang
     service:
@@ -44,10 +53,16 @@ projects:
       name:  bookhself
       cookie: bookshelf
       node: bookshelf@127.0.0.1
+    git:
+      uri: https://github.com/chef/oc_erchef.git
+      branch: master
+    build:
+      # higher concurrency appears to cause eunit failures...
+      max_concurrency: 1
   omnibus:
     type: omnibus
     git:
-      uri: git@github.com:chef/opscode-omnibus.git
+      uri: https://github.com/chef/opscode-omnibus.git
       branch: master
     components:
       # dest paths are relative to /opt/opscode/embedded for these components
