+++ dev/dvm/lib/dvm.rb
@@ -1,3 +1,4 @@
+require_relative "dvm/tools"
 require_relative "dvm/project"
 require_relative "dvm/application"
 require_relative "dvm/populate"
